+++ src/doc/changelog.md
@@ -1,3 +1,7 @@
+## v0.x.0
+
+- Add Japan GSI providers
+
 ## v0.1.0
 
 Initial version